Centralize sandbox PHP template fragments
