+++ packages/runtime-playground/src/browser-command-runners.ts
@@ -16,6 +16,7 @@ import { BROWSER_PROBE_CAPTURE_VALUES, BROWSER_PROBE_PERFORMANCE_INIT_SCRIPT, BR
 import { argValue, cleanWpCliOutput, commaListArg, durationArg, jsonArrayArg, strictBooleanArg, viewportArg } from "./commands.js"
 import { editorActionStepsFromArgs, editorOpenTargetFromArgs, type EditorActionStep } from "./editor-actions.js"
 import { bootstrapPhpCode } from "./php-bootstrap.js"
+import { phpBrowserWordPressDiagnosticsPlugin } from "./php-snippets.js"
 import { assertPlaygroundResponseOk, type PlaygroundRunResponse } from "./playground-command-errors.js"
 import type { PlaygroundCliServer } from "./preview-server.js"
 import type { Page } from "playwright"
@@ -1464,102 +1465,7 @@ interface BrowserWordPressDiagnosticsArtifact {
 
 const BROWSER_WORDPRESS_DIAGNOSTICS_LOG = "/wordpress/wp-content/wp-codebox-browser-diagnostics.jsonl"
 const BROWSER_WORDPRESS_DIAGNOSTICS_MU_PLUGIN = "/wordpress/wp-content/mu-plugins/000-wp-codebox-browser-diagnostics.php"
-const BROWSER_WORDPRESS_DIAGNOSTICS_PLUGIN = `<?php
-/**
- * Plugin Name: WP Codebox Browser Diagnostics
- */
-
-if ( ! defined( 'WPINC' ) ) {
-    return;
-}
-
-function wp_codebox_browser_diagnostics_write( array $record ): void {
-    file_put_contents( WP_CONTENT_DIR . '/wp-codebox-browser-diagnostics.jsonl', wp_json_encode( $record ) . "\n", FILE_APPEND | LOCK_EX );
-}
-
-function wp_codebox_browser_diagnostics_backtrace(): array {
-    $frames = debug_backtrace( DEBUG_BACKTRACE_IGNORE_ARGS, 12 );
-    $frames = array_slice( $frames, 2 );
-
-    return array_map(
-        static function ( array $frame ): array {
-            return array_filter(
-                array(
-                    'file'     => isset( $frame['file'] ) ? (string) $frame['file'] : null,
-                    'line'     => isset( $frame['line'] ) ? (int) $frame['line'] : null,
-                    'function' => isset( $frame['function'] ) ? (string) $frame['function'] : null,
-                    'class'    => isset( $frame['class'] ) ? (string) $frame['class'] : null,
-                    'type'     => isset( $frame['type'] ) ? (string) $frame['type'] : null,
-                ),
-                static fn ( $value ) => null !== $value && '' !== $value
-            );
-        },
-        $frames
-    );
-}
-
-add_filter(
-    'status_header',
-    static function ( string $status_header, int $code ): string {
-        if ( $code >= 500 && $code < 600 ) {
-            wp_codebox_browser_diagnostics_write(
-                array(
-                    'schema'         => 'wp-codebox/browser-wordpress-diagnostic-record/v1',
-                    'classification' => 'http-5xx-status',
-                    'severity'       => 'error',
-                    'status'         => $code,
-                    'statusHeader'   => $status_header,
-                    'requestUri'     => isset( $_SERVER['REQUEST_URI'] ) ? (string) $_SERVER['REQUEST_URI'] : '',
-                    'message'        => 'WordPress emitted a 5xx status header during browser navigation.',
-                    'backtrace'      => wp_codebox_browser_diagnostics_backtrace(),
-                    'capturedAt'     => gmdate( 'c' ),
-                )
-            );
-        }
-
-        return $status_header;
-    },
-    10,
-    2
-);
-
-register_shutdown_function(
-    static function (): void {
-        $error = error_get_last();
-        $fatal_types = array( E_ERROR, E_PARSE, E_CORE_ERROR, E_COMPILE_ERROR, E_USER_ERROR, E_RECOVERABLE_ERROR );
-        if ( is_array( $error ) && in_array( $error['type'] ?? null, $fatal_types, true ) ) {
-            wp_codebox_browser_diagnostics_write(
-                array(
-                    'schema'         => 'wp-codebox/browser-wordpress-diagnostic-record/v1',
-                    'classification' => 'php-fatal',
-                    'severity'       => 'error',
-                    'errorType'      => (int) ( $error['type'] ?? 0 ),
-                    'message'        => (string) ( $error['message'] ?? '' ),
-                    'file'           => (string) ( $error['file'] ?? '' ),
-                    'line'           => (int) ( $error['line'] ?? 0 ),
-                    'capturedAt'     => gmdate( 'c' ),
-                )
-            );
-            return;
-        }
-
-        $status = http_response_code();
-        if ( is_int( $status ) && $status >= 500 && $status < 600 ) {
-            wp_codebox_browser_diagnostics_write(
-                array(
-                    'schema'         => 'wp-codebox/browser-wordpress-diagnostic-record/v1',
-                    'classification' => 'http-response-code-5xx',
-                    'severity'       => 'error',
-                    'status'         => $status,
-                    'requestUri'     => isset( $_SERVER['REQUEST_URI'] ) ? (string) $_SERVER['REQUEST_URI'] : '',
-                    'message'        => 'Browser navigation finished with a 5xx HTTP response code and no PHP fatal.',
-                    'capturedAt'     => gmdate( 'c' ),
-                )
-            );
-        }
-    }
-);
-`
+const BROWSER_WORDPRESS_DIAGNOSTICS_PLUGIN = phpBrowserWordPressDiagnosticsPlugin()
 
 async function installBrowserWordPressDiagnostics(
   runPlaygroundCommand: ((command: string, server: PlaygroundCliServer, options: { code: string } | { scriptPath: string }) => Promise<PlaygroundRunResponse>) | undefined,
